lascou

+++ database/migrations/20230126170026-create-clientes.js
@@ -20,7 +20,7 @@ module.exports = {
         allowNull: false
       },
       tel: {
-        type: DataTypes.INTEGER(15),
+        type: DataTypes.INTEGER,
         allowNull: false
       },
       createdAt: {
